requirements.txt
praw
pymongo
transformers
torch
python-dotenv
redis
fastapi
uvicorn[standard]
pydantic
kafka-python
aiokafka
